+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/poolstatus/PendingChange.java
@@ -1,0 +1,19 @@
+package com.concordium.sdk.responses.poolstatus;
+
+import com.fasterxml.jackson.annotation.JsonSubTypes;
+import com.fasterxml.jackson.annotation.JsonTypeInfo;
+import lombok.Data;
+import lombok.Getter;
+import lombok.ToString;
+
+@Getter
+@Data
+@ToString
+@JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "pendingChangeType")
+@JsonSubTypes({
+        @JsonSubTypes.Type(value = PendingChangeNoChange.class, name = "NoChange"),
+        @JsonSubTypes.Type(value = PendingChangeRemovePool.class, name = "RemovePool"),
+        @JsonSubTypes.Type(value = PendingChangeReduceBakerCapital.class, name = "ReduceBakerCapital")
+})
+public abstract class PendingChange {
+}