+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/poolstatus/PendingChangeNoChange.java
@@ -1,0 +1,11 @@
+package com.concordium.sdk.responses.poolstatus;
+
+import lombok.Data;
+import lombok.Getter;
+import lombok.ToString;
+
+@Data
+@ToString
+@Getter
+public class PendingChangeNoChange extends PendingChange {
+}