+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/poolstatus/PendingChangeRemovePool.java
@@ -1,0 +1,18 @@
+package com.concordium.sdk.responses.poolstatus;
+
+import lombok.Data;
+import lombok.Getter;
+import lombok.ToString;
+
+import java.util.Date;
+
+@Getter
+@ToString
+@Data
+public class PendingChangeRemovePool extends PendingChange {
+
+    /**
+     * Effective time of the change.
+     */
+    private Date effectiveTime;
+}